refactor: 결과 메시지 생성 과정 원자 기능 분리
